feat: Desacoplado serviço Aws S3
