+++ TsundokuTraducoes.Helpers/DTOs/Utils/ParametroAwsServiceDto.cs
@@ -1,0 +1,9 @@
+﻿namespace TsundokuTraducoes.Helpers.DTOs.Utils;
+
+public class ParametroAwsServiceDto
+{
+    public bool AlterarImagem { get; set; }
+    public string NomeDiretorioImagens { get; set; }
+    public string NomeArquivoImagem { get; set; }
+    public byte[] ArrayByteImagem { get; set; }
+}
